Move new inside command.
